Add REST APIs methods to JavaScript templates
