+++ src/utilities/index.ts
@@ -2,5 +2,6 @@ export { getConfig } from './modules/config';
 export { getTemplateWithPartials } from './modules/partials';
 export { getCSSVariables } from './modules/css';
 export { randomId, flushPromise } from './modules/misc';
+export { getRestApis } from './modules/restApi';
 export * from './modules/predicates';
 export { logVersionToConsole } from './modules/logger';
